Extract fuzzy_match_tool_name into its own module to satisfy 500-LoC cap
The new separator-normalized strategy pushed validation.rs to 524 lines —
over the 500-LoC enforcement gate. Move all fuzzy-match logic
(fuzzy_match_tool_name + normalize_tool_name) into a sibling module
`fuzzy_match.rs` and re-import the entry point. Pure refactor, no
behavior change. Tests in tool_parser/tests/group_c.rs still exercise
the public path via validate_tool_calls, so the new private function in
fuzzy_match.rs is covered transitively.
